// web/src/shared/api/types.ts
export interface ActivityLogEvent {
  id: number;
  timestamp: string;
  user_id: number | null;
  username: string;
  location?: string;
  ip: string | null;
  event: ActivityLogEventTypeValue;
  module: ActivityLogModuleValue;
  device: string;
  description?: string;
}